src/newDIR/python/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
openai==1.12.0
anthropic==0.18.1
langchain==0.1.5
langchain-openai==0.0.5
langchain-anthropic==0.0.1
transformers==4.37.2
huggingface-hub==0.20.3
crewai==0.1.35
scikit-learn==1.4.0
xgboost==2.0.3
catboost==1.2.2
spacy==3.7.2
opencv-python==4.9.0.80
pymongo==4.6.1
redis==5.0.1
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
kafka-python==2.0.2
boto3==1.34.25
prometheus-client==0.19.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
sentry-sdk[fastapi]==1.40.0
pandas==2.2.0
numpy==1.26.3
